# reader-api/docs/swagger.yaml
basePath: /v1
definitions:
  handlers.Project:
    properties:
      createdAt:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      ocrData: {}
      updatedAt:
        type: string
    type: object
  handlers.Task:
    properties:
      createdAt:
        type: string
      id:
        type: string
      status:
        enum:
        - processing
        - error
        - completed
        - waiting
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  handlers.TaskWithProject:
    properties:
      createdAt:
        type: string
      id:
        type: string
      project:
        $ref: '#/definitions/handlers.Project'
      status:
        enum:
        - processing
        - error
        - completed
        - waiting
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for JP Reader application.
  title: JP Reader API
  version: "1.0"
paths:
  /healthz:
    get:
      consumes:
      - application/json
      description: Returns 200 if the server is ready
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Checks if the server is ready
      tags:
      - health
  /projects/all:
    get:
      consumes:
      - application/json
      description: Retrieves all projects from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Project'
            type: array
        "400":
          description: 'Couldn''t get projects: [error message]'
          schema:
            type: string
      summary: Get list of projects
      tags:
      - projects
  /tasks:
    get:
      consumes:
      - application/json
      description: Retrieves all tasks from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Task'
            type: array
        "400":
          description: 'Couldn''t get tasks: [error message]'
          schema:
            type: string
      summary: Get list of tasks
      tags:
      - tasks
  /tasks/projects:
    get:
      consumes:
      - application/json
      description: Retrieves all tasks along with their associated projects from the
        database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskWithProject'
            type: array
        "400":
          description: 'Couldn''t get tasks: [error message]'
          schema:
            type: string
      summary: Get list of tasks with projects
      tags:
      - tasks
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads multiple files, saves them to a directory, creates a task
        record in the database, and sends the task to the queue
      parameters:
      - description: Title of the task
        in: formData
        name: title
        required: true
        type: string
      - description: Files to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/plain
      responses:
        "200":
          description: Task has been queued
          schema:
            type: string
        "400":
          description: Title is required / No files uploaded / Error retrieving file
          schema:
            type: string
        "500":
          description: Error creating upload folder / Error saving file / Error copying
            file / Error creating task record / Error sending task to queue
          schema:
            type: string
      summary: Upload multiple files and queue task
      tags:
      - uploads
swagger: "2.0"